+++ crates/forge/src/lib.rs
@@ -33,6 +33,7 @@ pub const CAIRO_EDITION: &str = "2024_07";
 
 const MINIMAL_RUST_VERSION: Version = Version::new(1, 80, 1);
 const MINIMAL_SCARB_VERSION: Version = Version::new(2, 7, 0);
+const MINIMAL_RECOMMENDED_SCARB_VERSION: Version = Version::new(2, 8, 5);
 const MINIMAL_SCARB_VERSION_PREBUILT_PLUGIN: Version = Version::new(2, 10, 0);
 const MINIMAL_USC_VERSION: Version = Version::new(2, 0, 0);
 
@@ -277,6 +278,7 @@ fn check_requirements(output_on_success: bool) -> Result<()> {
         name: "Scarb".to_string(),
         command: RefCell::new(ScarbCommand::new().arg("--version").command()),
         minimal_version: MINIMAL_SCARB_VERSION,
+        minimal_recommended_version: Some(MINIMAL_RECOMMENDED_SCARB_VERSION),
         helper_text: "Follow instructions from https://docs.swmansion.com/scarb/download.html"
             .to_string(),
         version_parser: create_version_parser("Scarb", r"scarb (?<version>[0-9]+.[0-9]+.[0-9]+)"),
@@ -285,6 +287,7 @@ fn check_requirements(output_on_success: bool) -> Result<()> {
         name: "Universal Sierra Compiler".to_string(),
         command: RefCell::new(UniversalSierraCompilerCommand::new().arg("--version").command()),
         minimal_version: MINIMAL_USC_VERSION,
+        minimal_recommended_version: None,
         helper_text: "Reinstall `snforge` using the same installation method or follow instructions from https://foundry-rs.github.io/starknet-foundry/getting-started/installation.html#universal-sierra-compiler-update".to_string(),
         version_parser: create_version_parser(
             "Universal Sierra Compiler",
@@ -304,6 +307,7 @@ fn check_requirements(output_on_success: bool) -> Result<()> {
                 cmd
             }),
             minimal_version: MINIMAL_RUST_VERSION,
+            minimal_recommended_version: None,
             version_parser: create_version_parser(
                 "Rust",
                 r"rustc (?<version>[0-9]+.[0-9]+.[0-9]+)",